services/emotion_hse/requirements.txt
fastapi
uvicorn[standard]
pydantic
hsemotion-onnx
numpy
pillow